refactor: カテゴリ予算の変換処理を共通化
重複していた有効カテゴリ予算レスポンスのschemaとMap変換をdomain層へ集約する。
